# agentex/openapi.yaml
openapi: 3.1.0
info:
  title: Agentex API
  version: 0.1.0
paths:
...
  /tasks:
    get:
      tags:
      - Tasks
      summary: List Tasks
      description: List all tasks.
      operationId: list_tasks_tasks_get
      parameters:
      - name: agent_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
      - name: agent_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Name
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TaskStatus'
          - type: 'null'
          description: Filter tasks by status (e.g. RUNNING, COMPLETED).
          title: Status
        description: Filter tasks by status (e.g. RUNNING, COMPLETED).
      - name: task_metadata
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'JSON-encoded object used to filter tasks via JSONB containment.
            Example: {"created_by_user_id": "abc-123"}.'
          title: Task Metadata
        description: 'JSON-encoded object used to filter tasks via JSONB containment.
          Example: {"created_by_user_id": "abc-123"}.'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: page_number
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page Number
      - name: order_by
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Order By
      - name: order_direction
        in: query
        required: false
        schema:
          type: string
          default: desc
          title: Order Direction
      - name: relationships
        in: query
        required: false
        schema:
          type: array
          items:
            $ref: '#/components/schemas/TaskRelationships'
          title: Relationships
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
                title: Response List Tasks Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /tasks/{task_id}/export:
    get:
      tags:
      - task-retention
      summary: Export Task
      description: 'Build a self-contained snapshot of a task''s content surfaces.


        Returns the exact payload format that POST /rehydrate accepts, so

        export → clean → rehydrate is a round-trip-equivalent operation.'
      operationId: export_task_tasks__task_id__export_get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportTaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{task_id}/clean:
    post:
      tags:
      - task-retention
      summary: Clean Task
      description: 'Delete content-bearing rows for a stale task.


        Refuses on active tasks, in-flight workflows, or unprocessed events

        regardless of `force`. The `force=true` flag only bypasses the

        idle-threshold check.'
      operationId: clean_task_tasks__task_id__clean_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleanTaskRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanTaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /tasks/{task_id}/rehydrate:
    post:
      tags:
      - task-retention
      summary: Rehydrate Task
      description: 'Restore content-bearing rows from a snapshot.


        Refuses if the task isn''t currently in a cleaned state, or if any supplied

        message/state ID already exists in Mongo (catches double-rehydrate).'
      operationId: rehydrate_task_tasks__task_id__rehydrate_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RehydrateTaskRequest'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ACPType:
      type: string
      enum:
      - sync
      - async
      - agentic
      title: ACPType
    Agent:
      properties:
        id:
          type: string
          title: Id
          description: The unique identifier of the agent.
        name:
          type: string
          title: Name
          description: The unique name of the agent.
        description:
          type: string
          title: Description
          description: The description of the action.
        status:
          $ref: '#/components/schemas/AgentStatus'
          description: The status of the action, indicating if it's building, ready,
            failed, etc.
          default: Unknown
        acp_type:
          $ref: '#/components/schemas/ACPType'
          description: The type of the ACP Server (Either sync or async)
        status_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Reason
          description: The reason for the status of the action.
        created_at:
          type: string
          format: date-time
          title: Created At
          description: The timestamp when the agent was created
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: The timestamp when the agent was last updated
        registration_metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Registration Metadata
          description: The metadata for the agent's registration.
        registered_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Registered At
          description: The timestamp when the agent was last registered
        agent_input_type:
          anyOf:
          - $ref: '#/components/schemas/AgentInputType'
          - type: 'null'
          description: The type of input the agent expects.
        production_deployment_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Production Deployment Id
          description: ID of the current production deployment.
      type: object
      required:
      - id
      - name
      - description
      - acp_type
      - created_at
      - updated_at
      title: Agent
...
    AgentInputType:
      type: string
      enum:
      - text
      - json
      title: AgentInputType
...
    AgentStatus:
      type: string
      enum:
      - Ready
      - Failed
      - Unknown
      - Deleted
      - Unhealthy
      title: AgentStatus
...
    CleanTaskRequest:
      properties:
        force:
          type: boolean
          title: Force
          description: Skip the idle-threshold check. Active-workflow and unprocessed-events
            checks still apply. Admin use only.
          default: false
        idle_days:
          type: integer
          minimum: 1.0
          title: Idle Days
          description: Idle threshold in days (ignored when force=true).
          default: 7
      type: object
      title: CleanTaskRequest
    CleanTaskResponse:
      properties:
        task_id:
          type: string
          title: Task Id
        cleaned_at:
          type: string
          format: date-time
          title: Cleaned At
        messages_deleted:
          type: integer
          title: Messages Deleted
        task_states_deleted:
          type: integer
          title: Task States Deleted
        events_deleted:
          type: integer
          title: Events Deleted
      type: object
      required:
      - task_id
      - cleaned_at
      - messages_deleted
      - task_states_deleted
      - events_deleted
      title: CleanTaskResponse
...
    DataContentEntity:
      properties:
        type:
          type: string
          const: data
          title: Type
          description: The type of the message, in this case `data`.
          default: data
        author:
          $ref: '#/components/schemas/MessageAuthor'
          description: The role of the messages author, in this case `system`, `user`,
            `assistant`, or `tool`.
        style:
          $ref: '#/components/schemas/MessageStyle'
          description: The style of the message. This is used by the client to determine
            how to display the message.
          default: static
        data:
          additionalProperties: true
          type: object
          title: Data
          description: The contents of the data message.
      type: object
      required:
      - author
      - data
      title: DataContentEntity
...
    ExportTaskResponse:
      properties:
        task_id:
          type: string
          title: Task Id
        messages:
          items:
            $ref: '#/components/schemas/TaskMessageEntity'
          type: array
          title: Messages
        task_states:
          items:
            $ref: '#/components/schemas/StateEntity'
          type: array
          title: Task States
      type: object
      required:
      - task_id
      title: ExportTaskResponse
      description: Wire format mirrors the entity directly — schema parity is intentional.
...
    FileAttachmentEntity:
      properties:
        file_id:
          type: string
          title: File Id
          description: The unique ID of the attached file
        name:
          type: string
          title: Name
          description: The name of the file
        size:
          type: integer
          title: Size
          description: The size of the file in bytes
        type:
          type: string
          title: Type
          description: The MIME type or content type of the file
      type: object
      required:
      - file_id
      - name
      - size
      - type
      title: FileAttachmentEntity
      description: Represents a file attachment in messages.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageAuthor:
      type: string
      enum:
      - user
      - agent
      title: MessageAuthor
    MessageStyle:
      type: string
      enum:
      - static
      - active
      title: MessageStyle
...
    ReasoningContentEntity:
      properties:
        type:
          type: string
          const: reasoning
          title: Type
          description: The type of the message, in this case `reasoning`.
          default: reasoning
        author:
          $ref: '#/components/schemas/MessageAuthor'
          description: The role of the messages author, in this case `system`, `user`,
            `assistant`, or `tool`.
        style:
          $ref: '#/components/schemas/MessageStyle'
          description: The style of the message. This is used by the client to determine
            how to display the message.
          default: static
        summary:
          items:
            type: string
          type: array
          title: Summary
          description: A list of short reasoning summaries
        content:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Content
          description: The reasoning content or chain-of-thought text
      type: object
      required:
      - author
      - summary
      title: ReasoningContentEntity
...
    RehydrateTaskRequest:
      properties:
        task_id:
          type: string
          title: Task Id
        messages:
          items:
            $ref: '#/components/schemas/TaskMessageEntity'
          type: array
          title: Messages
        task_states:
          items:
            $ref: '#/components/schemas/StateEntity'
          type: array
          title: Task States
      type: object
      required:
      - task_id
      title: RehydrateTaskRequest
      description: Same shape as the export response — round-trip parity.
...
    StateEntity:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
          description: The task state's unique id
        task_id:
          type: string
          title: Task Id
          description: ID of the task this state belongs to. The combination of task_id
            and agent_id is globally unique.
        agent_id:
          type: string
          title: Agent Id
          description: ID of the agent this state belongs to. The combination of task_id
            and agent_id is globally unique.
        state:
          additionalProperties: true
          type: object
          title: State
          description: The state object that contains arbitrary data
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
          description: The timestamp when the state was created
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
          description: The timestamp when the state was last updated
      type: object
      required:
      - task_id
      - agent_id
      - state
      title: StateEntity
      description: 'Represents a state in the agent system. A state is associated
        uniquely with a task and an agent.


        This entity is used to store states in MongoDB, with each state

        associated with a specific task and agent. The combination of task_id and
        agent_id is globally unique.


        The state is a dictionary of arbitrary data.'
...
    TaskMessageEntity:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
          description: The task message's unique id
        task_id:
          type: string
          title: Task Id
          description: ID of the task this message belongs to
        content:
          oneOf:
          - $ref: '#/components/schemas/TextContentEntity'
          - $ref: '#/components/schemas/DataContentEntity'
          - $ref: '#/components/schemas/ToolRequestContentEntity'
          - $ref: '#/components/schemas/ToolResponseContentEntity'
          - $ref: '#/components/schemas/ReasoningContentEntity'
          title: Content
          description: The content of the message. This content is not OpenAI compatible.
            These are messages that are meant to be displayed to the user.
          discriminator:
            propertyName: type
            mapping:
              data: '#/components/schemas/DataContentEntity'
              reasoning: '#/components/schemas/ReasoningContentEntity'
              text: '#/components/schemas/TextContentEntity'
              tool_request: '#/components/schemas/ToolRequestContentEntity'
              tool_response: '#/components/schemas/ToolResponseContentEntity'
        streaming_status:
          anyOf:
          - type: string
            enum:
            - IN_PROGRESS
            - DONE
          - type: 'null'
          title: In case of streaming, this indicates whether the message is still
            being streamed or has been completed
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
          description: The timestamp when the message was created
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
          description: The timestamp when the message was last updated
      type: object
      required:
      - task_id
      - content
      title: TaskMessageEntity
      description: 'Represents a message in the agent system.


        This entity is used to store messages in MongoDB, with each message

        associated with a specific task.'
...
    TaskRelationships:
      type: string
      enum:
      - agents
      title: TaskRelationships
      description: Task relationships that can be loaded
    TaskResponse:
      properties:
        id:
          type: string
          title: Unique Task ID
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Unique name of the task
        status:
          anyOf:
          - $ref: '#/components/schemas/TaskStatus'
          - type: 'null'
          title: The current status of the task
        status_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: The reason for the current task status
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: The timestamp when the task was created
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: The timestamp when the task was last updated
        cleaned_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: The timestamp when the task's content was cleaned for retention compliance;
            null when active
        params:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Task parameters
        task_metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Task metadata
        agents:
          anyOf:
          - items:
              $ref: '#/components/schemas/Agent'
            type: array
          - type: 'null'
          title: Agents associated with this task (only populated when 'agent' view
            is requested)
      type: object
      required:
      - id
      title: TaskResponse
      description: Task response model with optional related data based on relationships
    TaskStatus:
      type: string
      enum:
      - CANCELED
      - COMPLETED
      - FAILED
      - RUNNING
      - TERMINATED
      - TIMED_OUT
      - DELETED
      title: TaskStatus
...
    TextContentEntity:
      properties:
        type:
          type: string
          const: text
          title: Type
          description: The type of the message, in this case `text`.
          default: text
        author:
          $ref: '#/components/schemas/MessageAuthor'
          description: The role of the messages author, in this case `system`, `user`,
            `assistant`, or `tool`.
        style:
          $ref: '#/components/schemas/MessageStyle'
          description: The style of the message. This is used by the client to determine
            how to display the message.
          default: static
        format:
          $ref: '#/components/schemas/TextFormat'
          description: The format of the message. This is used by the client to determine
            how to display the message.
          default: plain
        content:
          type: string
          title: Content
          description: The contents of the text message.
        attachments:
          anyOf:
          - items:
              $ref: '#/components/schemas/FileAttachmentEntity'
            type: array
          - type: 'null'
          title: Attachments
          description: Optional list of file attachments with structured metadata.
      type: object
      required:
      - author
      - content
      title: TextContentEntity
...
    TextFormat:
      type: string
      enum:
      - markdown
      - plain
      - code
      title: TextFormat
...
    ToolRequestContentEntity:
      properties:
        type:
          type: string
          const: tool_request
          title: Type
          description: The type of the message, in this case `tool_request`.
          default: tool_request
        author:
          $ref: '#/components/schemas/MessageAuthor'
          description: The role of the messages author, in this case `system`, `user`,
            `assistant`, or `tool`.
        style:
          $ref: '#/components/schemas/MessageStyle'
          description: The style of the message. This is used by the client to determine
            how to display the message.
          default: static
        tool_call_id:
          type: string
          title: Tool Call Id
          description: The ID of the tool call that is being requested.
        name:
          type: string
          title: Name
          description: The name of the tool that is being requested.
        arguments:
          additionalProperties: true
          type: object
          title: Arguments
          description: The arguments to the tool.
      type: object
      required:
      - author
      - tool_call_id
      - name
      - arguments
      title: ToolRequestContentEntity
...
    ToolResponseContentEntity:
      properties:
        type:
          type: string
          const: tool_response
          title: Type
          description: The type of the message, in this case `tool_response`.
          default: tool_response
        author:
          $ref: '#/components/schemas/MessageAuthor'
          description: The role of the messages author, in this case `system`, `user`,
            `assistant`, or `tool`.
        style:
          $ref: '#/components/schemas/MessageStyle'
          description: The style of the message. This is used by the client to determine
            how to display the message.
          default: static
        tool_call_id:
          type: string
          title: Tool Call Id
          description: The ID of the tool call that is being responded to.
        name:
          type: string
          title: Name
          description: The name of the tool that is being responded to.
        content:
          title: Content
          description: The result of the tool.
      type: object
      required:
      - author
      - tool_call_id
      - name
      - content
      title: ToolResponseContentEntity
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...